+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.4.0 (Aug 31, 2026)
+* Switched to OpenTofu.
+* Added optional `notification` connection to enable critical monitoring.
+* Added CPU utilization alarm configured via `var.alert_thresholds`.
+
 # 0.3.0 (Jun 19, 2026)
 * Upgraded `nullstone-io/ns` provider to `~> 0.11.0`.
 * Used `aws_tags` from `data.ns_workspace` to tag all resources via provider `default_tags`.